In this commit, I removed the login and register functionality as it was not required by the client for this project. Additionally, I connected the notification with the button for adding products to the cart.
Another change I made was the addition of a categories component. This component displays the available categories, and the plan is to have each category open a new page when clicked, showing only the products belonging to that category.

Furthermore, I made improvements to the styling and made certain components responsive.